Namespaces OrganizationType in Organization model, adds failing init for Organization model
