--- FullCircle/FullCircle/Models/Organization/OrganizationType.swift
@@ -1,18 +1,0 @@
-//
-//  OrganizationType.swift
-//  FullCircle
-//
-//  Created by Sunni Tang on 7/21/20.
-//  Copyright © 2020 FullCircle. All rights reserved.
-//
-
-import Foundation
-
-enum OrganizationType: String {
-    case movement
-    case nonprofit
-    case forProfit
-    case lobby
-    case academic
-    case individual
-}
